requirements.txt
# Core Dependencies for Cappuccino Trading System
# GPU-enabled PyTorch (CUDA 12.1)
torch>=2.6.0
torchvision>=0.21.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.1.0

# Optimization and Hyperparameter Tuning
optuna>=4.5.0
optuna-dashboard>=0.15.0

# Data Processing
pandas>=2.3.0
numpy>=1.26.0
pandas-flavor>=0.7.0

# Financial Data APIs
alpaca-trade-api>=3.2.0
yfinance>=0.2.0

# Econometric and Statistical Analysis
statsmodels>=0.14.0
arch>=6.2.0
linearmodels>=5.3
scipy>=1.11.0
scikit-learn>=1.3.0

# Time Series Analysis
pmdarima>=2.0.4
sktime>=0.25.0

# Technical Indicators
ta>=0.11.0
stockstats>=0.6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
tables>=3.9.1
python-dotenv>=1.0.0
pyyaml>=6.0

# Database and Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# API and Web
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Economic Data APIs
fredapi>=0.5.1

# Ollama client for sentiment analysis (if using local models)
ollama>=0.1.0

# Logging and Monitoring
tensorboard>=2.15.0
wandb>=0.16.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
